updated docstrings and readme

+++ plugins/operators/load_dimension.py
@@ -4,9 +4,12 @@ from airflow.utils.decorators import apply_defaults
 
 class LoadDimensionOperator(BaseOperator):
     """
-
-
-
+    Run SQL transformations to create dimensions tables
+    Input:
+        redshift_conn_id
+        table: target table
+        sql_source: sql code to be run
+        append mode: switching between insert/ truncate-insert pattern
     """
 
     ui_color = '#80BD9E'
@@ -19,26 +22,23 @@ class LoadDimensionOperator(BaseOperator):
 
     @apply_defaults
     def __init__(self,
-                 # Define your operators params (with defaults) here
-                 # Example:
-                 # conn_id = your-connection-name
                  redshift_conn_id="",
                  table="",
                  sql_source="",
+                 append_mode=True
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
-        # Map params here
-        # Example:
-        # self.conn_id = conn_id
         self.redshift_conn_id = redshift_conn_id
         self.table = table
         self.sql_source = sql_source
+        self.append_mode = append_mode
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        redshift.run("DELETE FROM {}".format(self.table))
+        if not self.append_mode:
+            redshift.run("DELETE FROM {}".format(self.table))
 
         formatted_sql = LoadDimensionOperator.insert_sql.format(
             self.table,
@@ -47,4 +47,4 @@ class LoadDimensionOperator(BaseOperator):
 
         redshift.run(formatted_sql)
 
-        self.log.info('LoadDimensionOperator not implemented yet')
+        self.log.info('LoadDimensionOperator implemented.')
